# Conflicts:
#	CHANGELOG.md

+++ src/services/Process.php
@@ -383,6 +383,9 @@ class Process extends Component
             }
 
             $enabledForSite[$element->siteId] = $attributeData['enabled'];
+
+            // Set the global status to true if it's enabled for *any* sites, or if already enabled.
+            $element->enabled = in_array(true, $enabledForSite) || $element->enabled;
             $element->setEnabledForSite($enabledForSite);
         }
 
@@ -392,7 +395,9 @@ class Process extends Component
                 $fieldValue = Plugin::$plugin->fields->parseField($feed, $element, $feedData, $fieldHandle, $fieldInfo);
 
                 if ($fieldValue !== null) {
-                    $fieldData[$fieldHandle] = $fieldValue;
+                    if ($feed['setEmptyValues'] === 1 || ($feed['setEmptyValues'] === 0 && !empty($fieldValue))) {
+                        $fieldData[$fieldHandle] = $fieldValue;
+                    }
                 }
             }
         }
@@ -527,22 +532,24 @@ class Process extends Component
             return;
         }
 
-        $elementsToDeleteDisable = array_diff($settings['existingElements'], $processedElementIds);
+        if ($processedElementIds) {
+            $elementsToDeleteDisable = array_diff($settings['existingElements'], $processedElementIds);
 
-        if ($elementsToDeleteDisable) {
-            if (DuplicateHelper::isDisable($feed)) {
-                $this->_service->disable($elementsToDeleteDisable);
-                $message = 'The following elements have been disabled: ' . Json::encode($elementsToDeleteDisable) . '.';
-            } elseif (DuplicateHelper::isDisableForSite($feed)) {
-                $this->_service->disableForSite($elementsToDeleteDisable);
-                $message = 'The following elements have been disabled for the target site: ' . Json::encode($elementsToDeleteDisable) . '.';
-            } else {
-                $this->_service->delete($elementsToDeleteDisable);
-                $message = 'The following elements have been deleted: ' . Json::encode($elementsToDeleteDisable) . '.';
+            if ($elementsToDeleteDisable) {
+                if (DuplicateHelper::isDisable($feed)) {
+                    $this->_service->disable($elementsToDeleteDisable);
+                    $message = 'The following elements have been disabled: ' . Json::encode($elementsToDeleteDisable) . '.';
+                } elseif (DuplicateHelper::isDisableForSite($feed)) {
+                    $this->_service->disableForSite($elementsToDeleteDisable);
+                    $message = 'The following elements have been disabled for the target site: ' . Json::encode($elementsToDeleteDisable) . '.';
+                } else {
+                    $this->_service->delete($elementsToDeleteDisable);
+                    $message = 'The following elements have been deleted: ' . Json::encode($elementsToDeleteDisable) . '.';
+                }
+
+                Plugin::info($message);
+                Plugin::debug($message);
             }
-
-            Plugin::info($message);
-            Plugin::debug($message);
         }
 
         // Log the total time taken to process the feed
@@ -551,7 +558,7 @@ class Process extends Component
 
         Plugin::$stepKey = null;
 
-        $message = 'Processing ' . count($processedElementIds) . ' elements finished in ' . $execution_time . 's';
+        $message = 'Processing ' . ($processedElementIds ? count($processedElementIds) : 0) . ' elements finished in ' . $execution_time . 's';
         Plugin::info($message);
         Plugin::debug($message);
